+++ custom_components/recalbox/frontend/recalbox-card.js
@@ -11,7 +11,10 @@ const TRANSLATIONS = {
       "shutdown": "Éteindre",
       "reboot": "Redémarrer",
       "screenshot": "Capture",
-      "stop": "Stop"
+      "pause": "Pause",
+      "save": "Enregistrer",
+      "load": "Restaurer",
+      "stop": "Quitter le jeu"
     },
     "footer": {
         "onHardware": "sur",
@@ -30,7 +33,10 @@ const TRANSLATIONS = {
       "shutdown": "Shutdown",
       "reboot": "Reboot",
       "screenshot": "Screenshot",
-      "stop": "Stop"
+      "stop": "Quit game",
+      "pause": "Pause",
+      "save": "Save",
+      "load": "Restore",
     },
     "footer": {
         "onHardware": "on",
@@ -84,9 +90,11 @@ class RecalboxCard extends HTMLElement {
             .game-preview { text-align: center; padding: 10px 0; margin: 10px -16px; }
             .game-preview img { max-width: 90%; border-radius: 4px; box-shadow: 0 2px 4px rgba(0,0,0,0.5); }
 
-            .card-actions { display: flex; gap: 8px; justify-content: center; padding: 12px; border-top: 1px solid var(--divider-color); border-radius: 0 0 12px 12px; }
+            .card-actions { display: flex; flex-wrap: wrap; gap: 8px; justify-content: center; padding: 12px; border-top: 1px solid var(--divider-color); border-radius: 0 0 12px 12px; }
             .action-button { display: flex; flex-direction: row; gap: 6px; border-radius: 20px; padding: 2px 12px; align-items: center; cursor: pointer; font-size: 10px; text-transform: uppercase; color: var(--primary-text-color); background-color: var(--chip-background-color); }
             .action-button ha-icon { color: var(--_leading-icon-color); margin-bottom: 4px; --mdc-icon-size: 18px; }
+            .game-icon-button { padding: 0 2px; }
+            .game-icon-button:hover { background-color: var(--chip-background-color); }
 
             .card-markdown-footer { padding: 24px 16px 16px; font-size: 0.8em; color: var(--secondary-text-color); line-height: 1.4; }
             .card-markdown-footer hr { border: 0; border-top: 1px solid var(--divider-color); margin: 16px 0; }
@@ -183,13 +191,17 @@ class RecalboxCard extends HTMLElement {
     if (isOn) {
       this.actions.style.display = "flex";
       this.actions.innerHTML = `
-        <div class="action-button" id="btn-power-off" ` + (showTurnOffButton ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:power"></ha-icon>${i18n.buttons.shutdown}</div>
-        <div class="action-button" id="btn-reboot" ` + (showRebootButton ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:restart"></ha-icon>${i18n.buttons.reboot}</div>
-        <div class="action-button" id="btn-snap" ` + ((showScreenshotButton && isAGameRunning) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:camera"></ha-icon>${i18n.buttons.screenshot}</div>
-        <div class="action-button" id="btn-pause" ` + ((showPauseGameButton && isAGameRunning) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:play-pause"></ha-icon>${i18n.buttons.pause}</div>
-        <div class="action-button" id="btn-save" ` + ((showSaveGameButton && isAGameRunning) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:content-save"></ha-icon>${i18n.buttons.save}</div>
-        <div class="action-button" id="btn-load" ` + ((showLoadGameButton && isAGameRunning) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:restore"></ha-icon>${i18n.buttons.load}</div>
-        <div class="action-button" id="btn-stop" ` + ((showQuitGameButton && isAGameRunning) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:location-exit"></ha-icon>${i18n.buttons.stop}</div>
+        <div class="action-button" id="btn-power-off" ` + ((showTurnOffButton) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:power"></ha-icon>${i18n.buttons.shutdown}</div>
+        <div class="action-button" id="btn-reboot" ` + ((showRebootButton) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:restart"></ha-icon>${i18n.buttons.reboot}</div>
+
+        <div class="action-button" ` + (isAGameRunning ? '' : 'style="display:none"')+ `>
+            ${i18n.game} &nbsp;
+            <div class="game-icon-button" title="${i18n.buttons.screenshot}" id="btn-snap" ` + ((showScreenshotButton) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:camera"></ha-icon></div>
+            <div class="game-icon-button" title="${i18n.buttons.pause}" id="btn-pause" ` + ((showPauseGameButton) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:play-pause"></ha-icon></div>
+            <div class="game-icon-button" title="${i18n.buttons.save}" id="btn-save" ` + ((showSaveGameButton) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:content-save"></ha-icon></div>
+            <div class="game-icon-button" title="${i18n.buttons.load}" id="btn-load" ` + ((showLoadGameButton) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:restore"></ha-icon></div>
+            <div class="game-icon-button" title="${i18n.buttons.stop}" id="btn-stop" ` + ((showQuitGameButton) ? '' : 'style="display:none"')+ `><ha-icon icon="mdi:stop"></ha-icon></div>
+        </div>
       `;
       this.actions.querySelector('#btn-power-off').onclick = () => hass.callService('recalbox', 'shutdown', { entity_id: entityId });
       this.actions.querySelector('#btn-reboot').onclick = () => hass.callService('recalbox', 'reboot', { entity_id: entityId });