+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Move ports configuration to an advanced section
 - Rename API ports fields to be more convenient (entities impacted)
 - Clean Recalbox entity attributes
+- Add a ping (optionnal) test on Recalbox creation
 
 
 ## v1.2.5 - 28/01/2026