+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 - Add Recalbox Card options to display or not buttons
 - Add Recalbox Card option to display Rom path (false by default)
 
+**Sauvegarder/Charger un state**
+- Ajout des fonctions de sauvegarde/chargement du dernier state
+- Exposition des services à Home Assistant
+- Ajout des intents par la voix/text pour Assist
+
 
 ## v1.2.6 - 28/01/2026